app/src/test/java/br/com/coupledev/kotlinflow/MainViewModelTest.kt
package br.com.coupledev.kotlinflow

import app.cash.turbine.test
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.launch
import kotlinx.coroutines.runBlocking
import org.junit.Assert.*

import org.junit.Before
import org.junit.Test

@ExperimentalCoroutinesApi
class MainViewModelTest {

    private lateinit var viewModel: MainViewModel
    private lateinit var testDispatchers: TestDispatchers

    @Before
    fun setUp() {
        testDispatchers = TestDispatchers()
        viewModel = MainViewModel(testDispatchers)
    }

    @Test
    fun `countDownFlow, properly count down from 10 to 0`() = runBlocking {
        viewModel.countDownFlow.test {
            for (i in 10 downTo 0) {
                testDispatchers.testDispatcher.advanceTimeBy(1000L)
                val emission = awaitItem()
                assertEquals(emission, i)
            }
            cancelAndConsumeRemainingEvents()
        }
    }

    @Test
    fun `squareNumber, number properly squared`() = runBlocking {
        val job = launch {
            viewModel.shareFlow.test {
                val emission = awaitItem()
                assertEquals(emission, 9)
            }
        }
        viewModel.squareNumber(3)
        job.join()
        job.cancel()
    }
}